telegram_bot/requirements.txt
aiogram
psycopg2-binary
matplotlib
pandas
